Move UDF-related stuff to '_helpers'.
