+++ google/cloud/bigquery/job.py
@@ -23,62 +23,10 @@ from google.cloud.bigquery.schema import SchemaField
 from google.cloud.bigquery.table import Table
 from google.cloud.bigquery.table import _build_schema_resource
 from google.cloud.bigquery.table import _parse_schema_resource
+from google.cloud.bigquery._helpers import UDFResourcesProperty
 from google.cloud.bigquery._helpers import _EnumProperty
 from google.cloud.bigquery._helpers import _TypedProperty
-
-
-class UDFResource(object):
-    """Describe a single user-defined function (UDF) resource.
-    :type udf_type: str
-    :param udf_type: the type of the resource ('inlineCode' or 'resourceUri')
-
-    :type value: str
-    :param value: the inline code or resource URI.
-
-    See
-    https://cloud.google.com/bigquery/user-defined-functions#api
-    """
-    def __init__(self, udf_type, value):
-        self.udf_type = udf_type
-        self.value = value
-
-    def __eq__(self, other):
-        return(
-            self.udf_type == other.udf_type and
-            self.value == other.value)
-
-
-def _build_udf_resources(resources):
-    """
-    :type resources: sequence of :class:`UDFResource`
-    :param resources: fields to be appended.
-
-    :rtype: mapping
-    :returns: a mapping describing userDefinedFunctionResources for the query.
-    """
-    udfs = []
-    for resource in resources:
-        udf = {resource.udf_type: resource.value}
-        udfs.append(udf)
-    return udfs
-
-
-class UDFResourcesProperty(object):
-    """Custom property type for :class:`QueryJob`.
-
-    Also used by :class:`~google.cloud.bigquery.query.Query`.
-    """
-    def __get__(self, instance, owner):
-        """Descriptor protocol:  accessor"""
-        if instance is None:
-            return self
-        return list(instance._udf_resources)
-
-    def __set__(self, instance, value):
-        """Descriptor protocol:  mutator"""
-        if not all(isinstance(u, UDFResource) for u in value):
-            raise ValueError("udf items must be UDFResource")
-        instance._udf_resources = tuple(value)
+from google.cloud.bigquery._helpers import _build_udf_resources
 
 
 class Compression(_EnumProperty):
@@ -952,7 +900,7 @@ class QueryJob(_AsyncJob):
 
     :type udf_resources: tuple
     :param udf_resources: An iterable of
-                        :class:`google.cloud.bigquery.job.UDFResource`
+                        :class:`google.cloud.bigquery._helpers.UDFResource`
                         (empty by default)
     """
     _JOB_TYPE = 'query'
